[DEBT-043-D15] Split analysis/correlation.py into focused modules
- Split analysis/correlation.py (448 LoC) into 3 files under 400 lines
- Created _correlation_models.py (98 LoC): CorrelationType, CorrelationResult, ArbitrageOpportunity, _is_priced
- Created _arbitrage.py (184 LoC): find_inverse_markets, find_inverse_market_groups, find_arbitrage_opportunities
- Reduced correlation.py to 274 LoC with CorrelationAnalyzer core
- CorrelationAnalyzer delegates arbitrage methods to module functions
- Re-exports preserve backwards-compatible public API
- Quality gates passed: pre-commit, pytest (1052 tests)
